Refactored code to handle io.Reader hanging problem.
I didn't realise that Readers are not supposed to return 0 bytes on read without an error. I also made the mistake of thinking I would get io.EOF when the HTTP message was fully received, instead of when the connection is terminated. So I've refactored the code such that we stop reading when the request has been fully parsed, not when there is no more data to read. This also helps simplify the code a bit and make the parsing state machine easier to understand.